Fix login
Loggin on using the login form was not possible, somehow no cookie was set on the server (/var/lib/php5/). To be honest, I have no idea as to why this fixes the problem but it works, don't have time to further investigate.

I am curious though as to how this could have worked in the first place.